# geotracking-synchronous-microservice/src/main/resources/openapi/geotracking-definitions.yaml
openapi: 3.0.1
info:
    title: Demo geo-tracking service REST API
    description: Geo tracking API
    version: 0.0.1

servers:
    - url: http://localhost:8081
      description: Generated server url

tags:
    - name: Tracking
      description: The tracking API


security:
    - bearerAuth: []

paths:
    /tracks:
        post:
            tags:
                - Tracking
            summary: Process a track
            operationId: processTrack
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Track'
            responses:
                "201":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TrackRef'
                "400":
                    description: Bad input parameters.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'


        get:
            tags:
                - Tracking
            summary: Get the track of a user in a time interval
            operationId: getTrackByParams
            parameters:
                -   name: user
                    in: query
                    description: User Id
                    required: true
                    schema:
                        type: string
                    example: W99999
                -   name: dateFrom
                    in: query
                    description: Track start date
                    required: true
                    schema:
                        type: string
                        format: date-time
                    example: "1975-07-05T07:30:00Z"
                -   name: dateTo
                    in: query
                    description: Track start date
                    required: true
                    schema:
                        type: string
                        format: date-time
                    example: "1975-07-05T07:30:00Z"

            responses:
                "200":
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/GeoPointResponse'
                "204":
                    description: Empty results.
                "400":
                    description: Bad input parameters.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'


    /tracks/lastPosition:
        get:
            tags:
                - Tracking
            summary: Get the last position of a user
            description: Get the last position of a user.
            operationId: getLastPosition
            parameters:
                -   name: testLazyTime
                    in: header
                    description: Time in seconds that the service is lazy.
                    required: false
                    schema:
                        type: integer
                        default: 0
                    example: 5

                -   name: user
                    in: query
                    description: User Id
                    required: true
                    schema:
                        type: string
                    example: X00005


            responses:
                "200":
                    description: Return one GeoPoint user item
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GeoPointResponse'
                "204":
                    description: Empty results.
                "400":
                    description: Bad input parameters.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'

components:
    schemas:

        Track:
            description: Is a sequence of consecutive points from a user
            type: object
            required:
                - deviceId
                - user
                - points
            properties:
                deviceId:
                    description: id device sent the track
                    type: string
                    example: 1111222XXXXXXX
                user:
                    description: User of the tracking
                    type: string
                    example: W99999
                points:
                    description: GeoPoins of the tracking. Minimum of 2 Points required
                    type: array
                    minItems: 2
                    items:
                        $ref: '#/components/schemas/Point'

        Point:
            description: geo point
            type: object
            required:
                - timestamp
                - point
            properties:
                timestamp:
                    description: timestamp of the GeoPoint.
                    type: string
                    format: date-time
                    example: 1975-07-05T07:30:00Z
                point:
                    $ref: '#/components/schemas/Geolocation'

        Geolocation:
            description: Geolocation coordinates
            type: object
            required:
                - longitude
                - latitude
            properties:
                longitude:
                    type: number
                    format: double
                    example: -8.255816
                latitude:
                    type: number
                    format: double
                    example: 43.459947

        TrackRef:
            description: Is a sequence of consecutive points from a user
            type: object
            properties:
                trackId:
                    description: id of the tracking
                    type: string
                    example: 6130dd549efc2a5c3edfd916


        GeoPointResponse:
            description: Return one GeoPoint user item
            type: object
            properties:
                id:
                    description: GeoPoint identifier
                    type: string
                    example: 641b94f35188c84d89373e99
                trackId:
                    description: GeoPoint identifier
                    type: string
                    example: 641b94f35188c84d89373e99
                user:
                    description: User of the tracking
                    type: string
                    example: W99999
                deviceId:
                    description: id device sent the track
                    type: string
                    example: 1111222XXXXXXX
                timestamp:
                    description: timestamp of the GeoPoint.
                    type: string
                    format: date-time
                    example: 1975-07-05T07:30:00Z
                point:
                    $ref: '#/components/schemas/Geolocation'


                      # TrackResponse:
                      #     description: Return one GeoPoint user item
                      #     type: object
                      #     properties:
                      #         user:
                      #             description: User of the tracking
                      #             type: string
                      #             example: W99999
                      #         points:
                      #             description: GeoPoins of the tracking
                      #             type: array
                      #             minItems: 2
                      #             items:
                    #                 $ref: '#/components/schemas/GeoPointResponse'


        # GeoPointResponse:
        #     description: geo point
        #     type: object
        #     properties:
        #         id:
        #             description: GeoPoint identifier
        #             type: string
        #             example: 641b94f35188c84d89373e99
        #         deviceId:
        #             description: id device sent the track
        #             type: string
        #             example: 1111222XXXXXXX
        #         timestamp:
        #             description: timestamp of the GeoPoint.
        #             type: string
        #             format: date-time
        #             example: 1975-07-05T07:30:00Z
        #         point:
        #             $ref: '#/components/schemas/Geolocation'


        # todo: eliminar
        # Audit:
        #     description: Audit data
        #     type: object
        #     properties:
        #         modifiedBy:
        #             description: User who made the last modification on the entity
        #             type: string
        #             example: W99999
        #         modifiedAt:
        #             description: Last modification date
        #             type: string
        #             format: date-time
        #             example: 1975-07-05T07:30:00Z
        #         createdBy:
        #             description: The user id who created de entity
        #             type: string
        #             example: W99999
        #         createdAt:
        #             description: Creation date
        #             type: string
        #             format: date-time
        #             example: 1975-07-05T07:30:00Z




        # Pagination todo:eliminar
        # Pagination:
        #     description: Pagination information. In the case that the payload of the response includes several pages of results
        #     type: object
        #     properties:
        #         page:
        #             type: integer
        #             description: Index of the current page
        #             example: 2
        #         size:
        #             type: integer
        #             description: Number of items per page
        #             example: 10
        #         totalElements:
        #             type: integer
        #             format: int64
        #             description: Total number of items found
        #             example: 25
        #         totalPages:
        #             type: integer
        #             description: Total number of pages found
        #             example: 3
        #         numberOfElements:
        #             type: integer
        #             description: Number of items on the current page
        #             example: 5
        #         next:
        #             description: Pagination key or index of the next page of results
        #             type: integer
        #             example: 2
        #         previous:
        #             description: Pagination key or index of the previous page of results
        #             type: integer
        #             example: 1
        #         sort:
        #             description: Pagination sorting criteria
        #             type: array
        #             items:
        #                 type: string
        #             example: ["date.desc","title.asc"]


        Error:
            description: error info
            required:
                - timestamp
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1975-07-05T07:30:00Z"
                status:
                    type: integer
                    format: int32
                error:
                    type: string
                message:
                    type: string
                path:
                    type: string